+++ matlab/src/functions/compute_dti_respairs.m
@@ -47,7 +47,7 @@ parfor fi = 1:length(data_folders)
     % Read in the DWI data
     fprintf('Loading DWI: %s (%s)\n', dw_file, data_folders{fi});
     tmp = load_nii( [input_folder dw_file] );
-    dw = flipdim(tmp.img,1);
+    dw = double(flipdim(tmp.img,1));
     [XSIZE,YSIZE,ZSIZE,~] = size(dw);
     hdr = tmp.hdr;
